requirements.txt
Flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
langchain>=0.1.0
openai>=1.6.1
langchain-openai>=0.0.2
requests>=2.31.0
python-dateutil>=2.8.2
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
langgraph>=0.0.15
typing-extensions>=4.8.0
pydantic>=2.5.1
gunicorn==21.2.0